+++ controller/controller.att.values.js
@@ -1,16 +1,11 @@
 import db from "../config/db.config.js";
+import Pagination from "../helpers/pagination.js";
 
 async function createAttValue(req, res) {
     try {
         const {name_uz, name_ru, atrebuts_id} = req.body
         if (!name_uz || !name_ru || !atrebuts_id) {
             const error = new Error('body not found')
-            error.status = 403
-            throw error
-        }
-        const [[attValue]] = await db.query("SELECT * FROM attributes_value WHERE name_uz = ?", name_uz)
-        if (attValue) {
-            const error = new Error('there is an attribute')
             error.status = 403
             throw error
         }
@@ -24,13 +19,14 @@ async function createAttValue(req, res) {
 
 async function getAttValues(req, res) {
     try {
-        const [attValues] = await db.query("SELECT * FROM attributes_value")
-        if (!attValues) {
-            const error = new Error("attribute not fount")
-            error.status = 402
-            throw error
-        }
-        res.json(attValues)
+        const { page, limit } = req.query;
+        // Fetch the total count of records
+        const [[{ "COUNT(*)": totalItems }]] = await db.query("SELECT COUNT(*) FROM attributes_value");
+        // Create a Pagination object
+        const paginations = new Pagination(totalItems, page, limit);
+        // Fetch the paginated addresses
+        const [attributes_value] = await db.query("SELECT * FROM attributes_value LIMIT ? OFFSET ?", [paginations.limit, paginations.offset]);
+        res.json({ attributes_value, paginations }.attributes_value);
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -76,8 +72,20 @@ async function updateAttValue(req, res) {
             error.status = 402
             throw error
         }
-        await db.query("UPDATE attributes_value SET ? WHERE id = ?", [body, id])
-        res.json("updated attributes_value id = "+ id)
+
+        if (req.role === "admin") {
+            await db.query("UPDATE attributes_value SET ? WHERE id = ?", [body, id])
+            res.json("updated attributes_value id = "+ id)
+            return
+        }
+        if (req.id === attValue.user_id) {
+            await db.query("UPDATE attributes_value SET ? WHERE id = ?", [body, id])
+            res.json("updated attributes_value id = "+ id)
+            return
+        }
+        const error = new Error("you are not the owner of this address")
+        error.status = 403
+        throw error
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -91,8 +99,20 @@ async function deleteAttValue(req, res) {
             error.status = 402
             throw error
         }
-        await db.query("DELETE FROM attributes_value WHERE id = ?", id)
-        res.json("delete attributes_value id = "+ id)
+
+        if (req.role === "admin") {
+            await db.query("DELETE FROM attributes_value WHERE id = ?", id)
+            res.json("delete attributes_value id = "+ id)
+            return
+        }
+        if (req.id === attValue.user_id) {
+            await db.query("DELETE FROM attributes_value WHERE id = ?", id)
+            res.json("delete attributes_value id = "+ id)
+            return
+        }
+        const error = new Error("you are not the owner of this address")
+        error.status = 403
+        throw error
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -107,7 +127,7 @@ export {
 }
 
 
-///////// test attributes
+///////// test attributes_value
 
 // {
 //     "name_uz": "Assalomu aleykum nima deyishni bilmadim null deyman",
