+++ controller/controller.cart.js
@@ -1,18 +1,19 @@
 import db from "../config/db.config.js";
+import Pagination from "../helpers/pagination.js";
 
 async function createCart(req, res) {
     try {
         const {user_id, product_id, count} = req.body
         if (!user_id || !product_id || !count) {
             const error = new Error('body not found')
-            error.status = 402
+            error.status = 400
             throw error
         }
         const [[cart]] = await db.query("SELECT * FROM cart WHERE user_id = ?", user_id)
         if (cart) {
             if (cart.product_id === product_id) {
                 const error = new Error('already have a cart')
-                error.status = 402
+                error.status = 400
                 throw error
             }
         }
@@ -26,13 +27,14 @@ async function createCart(req, res) {
 
 async function getCarts(req, res) {
     try {
-        const [cart] = await db.query("SELECT * FROM cart")
-        if (!cart) {
-            const error = new Error("cart not fount")
-            error.status = 402
-            throw error
-        }
-        res.json(cart)
+        const { page, limit } = req.query;
+        // Fetch the total count of records
+        const [[{ "COUNT(*)": totalItems }]] = await db.query("SELECT COUNT(*) FROM cart");
+        // Create a Pagination object
+        const paginations = new Pagination(totalItems, page, limit);
+        // Fetch the paginated cart
+        const [cart] = await db.query("SELECT * FROM cart LIMIT ? OFFSET ?", [paginations.limit, paginations.offset]);
+        res.json({ cart, paginations }.cart);
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -83,13 +85,13 @@ async function updateCart(req, res) {
             res.json("updated cart id = "+ id)
             return
         }
-        if (req.id === adress.user_id) {
+        if (req.id === cart.user_id) {
             await db.query("UPDATE cart SET ? WHERE id = ?", [body, id])
             res.json("updated cart id = "+ id)
             return
         }
         const error = new Error("you are not the owner of this account")
-        error.status = 403
+        error.status = 400
         throw error
     } catch (error) {
         res.status(error.status).json({error:error.message})
@@ -104,8 +106,20 @@ async function deleteCart(req, res) {
             error.status = 402
             throw error
         }
-        await db.query("DELETE FROM cart WHERE id = ?", id)
-        res.json("updated cart id = "+ id)
+
+        if (req.role === "admin") {
+            await db.query("DELETE FROM cart WHERE id = ?", id)
+            res.json("updated cart id = "+ id)
+            return
+        }
+        if (req.id === cart.user_id) {
+            await db.query("DELETE FROM cart WHERE id = ?", id)
+            res.json("delete cart id = "+ id)
+            return
+        }
+        const error = new Error("you are not the owner of this account")
+        error.status = 400
+        throw error
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
