+++ controller/controller.orders.js
@@ -1,17 +1,12 @@
 import db from "../config/db.config.js";
+import Pagination from "../helpers/pagination.js";
 
 async function createOrder(req, res) {
     try {
         const {adress_id, user_id, product, delevery_id, count, status} = req.body
         if (!adress_id || !user_id || !product || !delevery_id || !count || !status) {
             const error = new Error('body not found')
-            error.status = 403
-            throw error
-        }
-        const [[orders]] = await db.query("SELECT * FROM orders WHERE product = ?", product)
-        if (orders) {
-            const error = new Error('already have a orders')
-            error.status = 403
+            error.status = 400
             throw error
         }
 
@@ -24,13 +19,14 @@ async function createOrder(req, res) {
 
 async function getOrders(req, res) {
     try {
-        const [orders] = await db.query("SELECT * FROM orders")
-        if (!orders) {
-            const error = new Error("orders not fount")
-            error.status = 402
-            throw error
-        }
-        res.json(orders)
+        const { page, limit } = req.query;
+        // Fetch the total count of records
+        const [[{ "COUNT(*)": totalItems }]] = await db.query("SELECT COUNT(*) FROM orders");
+        // Create a Pagination object
+        const paginations = new Pagination(totalItems, page, limit);
+        // Fetch the paginated addresses
+        const [orders] = await db.query("SELECT * FROM orders LIMIT ? OFFSET ?", [paginations.limit, paginations.offset]);
+        res.json({ orders, paginations }.orders);
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -50,17 +46,7 @@ async function getOrder(req, res) {
             error.status = 402
             throw error
         }
-        if (req.role === "admin") {
-            res.json(orders)
-            return
-        }
-        if (orders.user_id === req.id) {
-            res.json(orders)
-            return
-        }
-        const error = new Error("this is not your order")
-        error.status = 404
-        throw error
+        res.json(orders)
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -86,8 +72,20 @@ async function updateOrder(req, res) {
             error.status = 402
             throw error
         }
-        await db.query("UPDATE orders SET ? WHERE id = ?", [body, id])
-        res.json("updated orders id = "+ id)
+
+        if (req.role === "admin") {
+            await db.query("UPDATE orders SET ? WHERE id = ?", [body, id])
+            res.json("updated orders id = "+ id)
+            return
+        }
+        if (req.id === orders.user_id) {
+            await db.query("UPDATE orders SET ? WHERE id = ?", [body, id])
+            res.json("updated orders id = "+ id)
+            return
+        }
+        const error = new Error("you are not the owner of this address")
+        error.status = 403
+        throw error
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -101,8 +99,20 @@ async function deleteOrder(req, res) {
             error.status = 402
             throw error
         }
-        await db.query("DELETE FROM orders WHERE id = ?", id)
-        res.json("delete orders id = "+ id)
+
+        if (req.role === "admin") {
+            await db.query("DELETE FROM orders WHERE id = ?", id)
+            res.json("delete orders id = "+ id)
+            return
+        }
+        if (req.id === orders.user_id) {
+            await db.query("DELETE FROM orders WHERE id = ?", id)
+            res.json("delete orders id = "+ id)
+            return
+        }
+        const error = new Error("you are not the owner of this address")
+        error.status = 403
+        throw error
     } catch (error) {
         res.status(error.status).json({error:error.message})
     }
@@ -110,18 +120,18 @@ async function deleteOrder(req, res) {
 
 export {
     createOrder,
-    getOrders,
     getOrder,
+    getOrders,
     updateOrder,
     deleteOrder
 }
 
-////////// test Order
+////////// test Orders
 
 // {
 //     "adress_id": 1,
 //     "user_id": 1,
-//     "product": [1,2],
+//     "product": 1,
 //     "delevery_id": 1,
 //     "count": 2,
 //     "status": "packing"
